rename ZeBits package and zeclasses
