+++ zeapp/app/src/main/java/de/berlindroid/zeapp/ui/ImageGenerationEditorDialog.kt
@@ -28,10 +28,10 @@ import de.berlindroid.zeapp.OPENAI_API_KEY
 import de.berlindroid.zeapp.PAGE_HEIGHT
 import de.berlindroid.zeapp.PAGE_WIDTH
 import de.berlindroid.zeapp.R
-import de.berlindroid.zeapp.bits.copy
-import de.berlindroid.zeapp.bits.cropPageFromCenter
-import de.berlindroid.zeapp.bits.isBinary
-import de.berlindroid.zeapp.bits.scaleIfNeeded
+import de.berlindroid.zeapp.zebits.copy
+import de.berlindroid.zeapp.zebits.cropPageFromCenter
+import de.berlindroid.zeapp.zebits.isBinary
+import de.berlindroid.zeapp.zebits.scaleIfNeeded
 import de.berlindroid.zeapp.vm.ZeBadgeViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
