+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zebits/ZeFloydSteinberg.kt
@@ -1,4 +1,4 @@
-package de.berlindroid.zeapp.bits
+package de.berlindroid.zeapp.zebits
 
 import android.graphics.Bitmap
 import android.graphics.Color
@@ -6,21 +6,24 @@ import java.nio.IntBuffer
 import kotlin.math.roundToInt
 
 /**
- * Attempt on a static positional pattern dithering algorithm.
+ * Floyd-Steinberg Dithering
  *
- * This will convert a color bitmap into a black and white image following a set amount of static
- * positional 3x3 patterns.
+ * Use this extension to create a binary bitmap, dithered following the Floyd-Steinberg algorithm.
+ *
+ * @see [Wikpedia](https://en.wikipedia.org/wiki/Floyd%E2%80%93Steinberg_dithering)
  */
-fun Bitmap.ditherPositional(): Bitmap {
+fun Bitmap.ditherFloydSteinberg(): Bitmap {
     val outputBitmap = grayscale()
 
+    // create new buffer, storing only one value, it's green,
+    // so the pixel color value can be temporarily bigger then 255
     val buffer = IntBuffer.allocate(width * height)
     outputBitmap.copyPixelsToBuffer(buffer)
     buffer.rewind()
 
     buffer.map { Color.green(it) }
 
-    // loop through all pixels. propagating the thresholding to specific pixel only error
+    // loop through all pixels. propagating the thresholding error
     for (y in 0 until height) {
         for (x in 0 until width) {
             val old = buffer[x + y * width]
@@ -28,11 +31,10 @@ fun Bitmap.ditherPositional(): Bitmap {
             val error = old - new
             buffer.put(x + y * width, new)
 
-            // select weight according to the index of the pixel
-            val weights = if ((x + y) % 2 == 0) EVEN_NEIGHBOR_WEIGHTS else ODD_NEIGHBOR_WEIGHTS
-            weights.forEachIndexed { index, weight ->
-                val i = index % 2 - 1
-                val j = index / 2
+            // distribute the error to the neighboring pixels
+            FLOYD_STEINBERG_NEIGHBOR_WEIGHTS.forEachIndexed { index, weight ->
+                val i = index % 3 - 1
+                val j = index / 3
 
                 if (weight > 0.0f) {
                     if (x + i in 0 until width) {
@@ -56,12 +58,7 @@ fun Bitmap.ditherPositional(): Bitmap {
     return outputBitmap
 }
 
-private val EVEN_NEIGHBOR_WEIGHTS = arrayOf(
-    0 / 4.0f, /**/0.0f, 0 / 4.0f,
-    2 / 4.0f, 0 / 4.0f, 2 / 4.0f,
-)
-
-private val ODD_NEIGHBOR_WEIGHTS = arrayOf(
-    0 / 4.0f, /**/0.0f, 2 / 4.0f,
-    0 / 4.0f, 2 / 4.0f, 0 / 4.0f,
+private val FLOYD_STEINBERG_NEIGHBOR_WEIGHTS = arrayOf(
+    0 / 16.0f, /**/ 0.0f, 7 / 16.0f,
+    3 / 16.0f, 5 / 16.0f, 1 / 16.0f
 )
